chore(scap): move api compatibility checks to libscap
We are about to use the checks for all engines (if they don't
support API versions, the check will be a no-op) so we cannot
provide the functions in a Linux-only library.
